requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
langchain>=0.0.300
langchain-community>=0.0.10
langchain-core>=0.0.10
langchain-huggingface>=0.0.5
llama-cpp-python>=0.2.0
pydantic>=2.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.18
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
transformers>=4.36.0
accelerate>=0.25.0
pypdf>=3.17.0
python-docx>=0.8.11
beautifulsoup4>=4.12.2
torch
torchvision
torchaudio 